MainPage для SportLineApp
