+++ src/main/java/com/sportline/SportLineApp.java
@@ -1,12 +1,12 @@
-package com.myspring;
+package com.sportline;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MySpringApplication {
+public class SportLineApp {
 
     public static void main(String[] args) {
-        SpringApplication.run(MySpringApplication.class, args);
+        SpringApplication.run(SportLineApp.class, args);
     }
 }
