+++ src/main/java/com/sportline/controller/BlogController.java
@@ -1,0 +1,5 @@
+package com.sportline.controller;
+
+public class BlogController {
+
+}
