+++ src/main/java/com/sportline/controller/HomeController.java
@@ -1,0 +1,17 @@
+package com.sportline.controller;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@AllArgsConstructor
+@RequestMapping("sportline")
+public class HomeController {
+
+    @GetMapping("")
+    public String getHomePage() {
+        return "sportline/main/home";
+    }
+}
