+++ src/main/java/com/sportline/controller/ProductController.java
@@ -1,0 +1,5 @@
+package com.sportline.controller;
+
+public class ProductController {
+
+}
