+++ src/main/java/com/sportline/controllers/AuthorsController.java
@@ -1,7 +1,7 @@
-package com.myspring.controllers;
+package com.sportline.controllers;
 
-import com.myspring.models.Author;
-import com.myspring.services.AuthorsService;
+import com.sportline.models.Author;
+import com.sportline.services.AuthorsService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
