+++ src/main/java/com/sportline/models/Author.java
@@ -1,10 +1,9 @@
-package com.myspring.models;
+package com.sportline.models;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 import java.util.List;
 
@@ -16,14 +15,11 @@ import java.util.List;
 public class Author {
 
     @Id
-    @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "full_name")
     private String fullName;
 
-    @ToString.Exclude
     @OneToMany(mappedBy = "author", cascade = CascadeType.REMOVE, fetch = FetchType.EAGER)
     private List<NewsCard> newsCards;
 }
