+++ src/main/java/com/sportline/models/NewsCard.java
@@ -1,0 +1,36 @@
+package com.sportline.models;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "newscards")
+public class NewsCard {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String header;
+
+    private String content;
+
+    @DateTimeFormat(pattern = "dd-MM-yyyy HH:mm:ss")
+    @Temporal(TemporalType.TIMESTAMP)
+    private LocalDateTime creationDate;
+
+    private Integer ratingPoints;
+
+    @ManyToOne
+    @JoinColumn(name = "author_id")
+    private Author author;
+}
