+++ src/main/java/com/sportline/repositories/DevicesRepository.java
@@ -1,6 +1,6 @@
-package com.myspring.repositories;
+package com.sportline.repositories;
 
-import com.myspring.models.NewsCard;
+import com.sportline.models.Device;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +8,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsCardsRepository extends JpaRepository<NewsCard, Long>, PagingAndSortingRepository<NewsCard, Long> {
+public interface DevicesRepository extends PagingAndSortingRepository<Device, Long>, JpaRepository<Device, Long> {
 
-    Page<NewsCard> findAllByOrderByIdDesc(Pageable pageable);
+    Page<Device> findAllByOrderByIdDesc(Pageable pageable);
 }
