+++ src/main/java/com/sportline/repositories/NewsCardsRepository.java
@@ -1,6 +1,6 @@
-package com.myspring.repositories;
+package com.sportline.repositories;
 
-import com.myspring.models.Device;
+import com.sportline.models.NewsCard;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +8,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DevicesRepository extends PagingAndSortingRepository<Device, Long>, JpaRepository<Device, Long> {
+public interface NewsCardsRepository extends JpaRepository<NewsCard, Long>, PagingAndSortingRepository<NewsCard, Long> {
 
-    Page<Device> findAllByOrderByIdDesc(Pageable pageable);
+    Page<NewsCard> findAllByOrderByIdDesc(Pageable pageable);
 }
