+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8090
+server.port=8099
 spring.mvc.hiddenmethod.filter.enabled=true
 
 spring.datasource.username=postgres
@@ -6,3 +6,4 @@ spring.datasource.password=3gor
 spring.datasource.url=jdbc:postgresql://localhost:5432/myspringmvc
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy
