+++ src/main/resources/static/css/light-theme.css
@@ -1,16 +1,16 @@
 @media (prefers-color-scheme: light) {
     body {
         font-family: 'Consolas L', sans-serif;
-        color: #303F30;
+        color: var(--m-dark-color);
         font-size: 2vh;
-        background-color: #EEEDEE;
+        background-color: var(--m-light-color);
     }
 
     .container, footer {
         width: 50vw;
         margin: 0 auto;
         padding: 2vw;
-        background-color: #DDDFDD;
+        background-color: var(--light-color);
     }
 
     footer {
@@ -19,25 +19,25 @@
     }
 
     a {
-        color: #2ECC71;
+        color: var(--secondary-color);
         text-decoration: none;
     }
 
     a:hover {
-        color: #FFF;
-        background-color: #2ECC71;
+        color: var(--light-color);
+        background-color: var(--secondary-color);
         text-decoration: none;
     }
 
     a:active {
-        color: #2ECC71;
-        background-color: #F5F5F5;
+        color: var(--secondary-color);
+        background-color: var(--m-light-color);
         text-decoration: none;
     }
 
     header a {
         padding: 0 5vw;
         font-size: 4vh;
-        background-color: #DDDFDD;
+        background-color: var(--m-light-color);
     }
 }
