+++ src/main/resources/static/css/tables.css
@@ -6,7 +6,7 @@ table {
 
 th, td {
     padding: 10px;
-    border-bottom: 1px solid #2ECC71;
+    border-bottom: 1px solid var(--accent-color);
 }
 
 th {
