+++ src/main/resources/templates/smart_home/list_authors.html
@@ -4,7 +4,7 @@
     <section>
         <h2>Авторы блога</h2>
         <div class="card-container">
-            <!--/*@thymesVar id="author" type="com.myspring.models.Author"*/-->
+            <!--/*@thymesVar id="author" type="com.sportline.models.Author"*/-->
             <div class="card" th:each="author : ${authors.content}">
                 <h3 th:text="*{author.fullName}"></h3>
                 <a th:href="@{/smarthome/authors/{id}(id=*{author.id})}" class="card-button">Подробнее</a>
