+++ src/main/resources/templates/sportline/main/footer.html
@@ -1,0 +1,36 @@
+</div>
+<footer>
+    <div>
+        <h3>Контакты</h3>
+        <a href="mailto:info@sportline.com">info@sportline.com</a>
+        <a href="tel:+79123456789">+7(912)345-67-89</a>
+    </div>
+    <div>
+        <h3>Ссылки</h3>
+        <div>
+            <a href="">Главная</a>
+            <a href="">Товары</a>
+            <a href="">Новости и Блог</a>
+            <a href="">Спецпредложения</a>
+        </div>
+    </div>
+    <div>
+        <h3>Социальные сети</h3>
+            <a href="vk.com/sportline">ВКонтакте</a>
+            <a href="t.me/sportline">Telegram</a>
+    </div>
+
+    <div style="position:relative;overflow:hidden;"><a
+            href="https://yandex.ru/maps/org/belgorodskiy_gosudarstvenny_natsionalny_issledovatelskiy_universitet/1235860900/?utm_medium=mapframe&utm_source=maps"
+            style="color:#eee;font-size:12px;position:absolute;top:0;">СпортЛайн</a><a
+            href="https://yandex.ru/maps/4/belgorod/category/university/184106140/?utm_medium=mapframe&utm_source=maps"
+            style="color:#eee;font-size:12px;position:absolute;top:14px;">СпортЛайн в Белгороде</a>
+        <iframe src="https://yandex.ru/map-widget/v1/?ll=36.576416%2C50.593562&mode=search&oid=1235860900&ol=biz&z=17"
+                width="720" height="400" allowfullscreen="allowfullscreen" style="position:relative;"></iframe>
+    </div>
+
+    <p>© 2024 СпортЛайн. Игорь Головков.</p>
+</footer>
+
+</body>
+</html>
