+++ src/main/resources/templates/sportline/offers/item_manage.html
@@ -1,0 +1,5 @@
+<div th:replace="~{/sportline/main/header.html}"></div>
+
+Добро пожаловать
+
+<div th:replace="~{/sportline/main/footer.html}"></div>
